Smooth scrolling by R Bibby 1987
The code is to smooth scroll text accross screen on a zx spectrum. It was a type in program in Popular computer weekly 1987, credited to R Bibby. I retyped it in from screen photo of the page of that magazine, and had to debug a few bits as I got some typos. One nicve thing is the machine code in the data statements are read in and uses a checksum on the end of each line of data.
The output is from Speccy5.6 emulator, using zx pritner. To view this file correctly you have to load into an editor such as notpad++ or a word processor and zoom out. each # character is a "pixel" of a bigger text character, and a full stop means no dot was "printed" on that lcoation. I tried to find a better way of uploading it.